Move the post score modal to the game component screen - Tucker
