+++ components/GameComponent.js
@@ -1,12 +1,13 @@
 import React from 'react';
 import { useEffect, useState } from 'react';
 import { StyleSheet, TouchableOpacity, View } from 'react-native';
-import { Text, useTheme } from 'react-native-paper';
+import { Modal, Text, TextInput, useTheme, HelperText} from 'react-native-paper';
 import { useProblems } from '../hooks/useProblems';
 import { GameResultComponent } from './GameResultComponent';
 import { GameStartComponent } from './GameStartComponent';
 import ProblemButtonComponent from './ProblemButtonComponent';
 import { runTransaction } from '../Utilities/dbUtils';
+import { createScoreObj, validateAlphaString } from '../Utilities/utils';
 
 export default function GameComponent({ navigation, gameType, startTime = 30, db }) {
   const [timer, setTimer] = useState(startTime)
@@ -16,6 +17,9 @@ export default function GameComponent({ navigation, gameType, startTime = 30, db
   const [startGame, setStartGame] = useState(false)
   const [problems, createNewProblems] = useProblems()
   const [savedScore, setSavedScore] = useState(undefined)
+  const [openPostScore, setOpenPostScore] = useState(false)
+  const [name, setName] = useState('')
+  const [errorMsg, setErrorMsg] = useState('')
 
   useEffect(() => {
     if (!startGame) {
@@ -37,7 +41,7 @@ export default function GameComponent({ navigation, gameType, startTime = 30, db
   const finishGame = async () => {
     const savedScore = await runTransaction(db, `INSERT INTO SCORE (GAME_TYPE, SCORE, DATE_PLAYED) VALUES ("${gameType}", ${score}, "${"2023-04-12"}") RETURNING *;`)
     setSavedScore(savedScore[0]);
-    setGameOver(true)
+    setOpenPostScore(true)
   }
 
   const theme = useTheme();
@@ -88,7 +92,7 @@ export default function GameComponent({ navigation, gameType, startTime = 30, db
       paddingVertical: 15,
       paddingHorizontal: 20,
       borderRadius: 15,
-      width: '100%',
+      margin: '3%'
     },
     buttonLabel: {
       fontSize: 35,
@@ -99,6 +103,38 @@ export default function GameComponent({ navigation, gameType, startTime = 30, db
       fontSize: 35,
       fontWeight: 'bold',
       textAlign: 'center'
+    },
+    postScoreContainer: {
+      margin: '5%',
+      backgroundColor: theme.colors.tertiary,
+      height: '40%',
+      width: '90%',
+      borderRadius: 5,
+    },
+    modalTitleRow: {
+      display: 'flex',
+      flexDirection: 'row',
+      borderBottomWidth: 3,
+      borderBottomColor: theme.colors.primary,
+      alignItems: 'center',
+      width: '100%',
+    },
+    modalTitle: {
+      flex: 1,
+      fontSize: 24,
+      borderBottomColor: theme.colors.primary,
+      margin: 10
+    },
+    formContainer: {
+      flex: 1,
+      display: 'flex',
+    },
+    formElement: {
+      marginHorizontal: '2%',
+      marginVertical: '2%'
+    },
+    nameInputContainer: {
+      flex: 1,
     }
   });
 
@@ -137,6 +173,43 @@ export default function GameComponent({ navigation, gameType, startTime = 30, db
     setStartGame(true)
   }
 
+  const handlePostScore = () => {
+    if (name.length > 25) {
+      setErrorMsg('Name must be less than 25 letters.')
+      return
+    }
+    if (!validateAlphaString(name)) {
+      setErrorMsg('Name must be letters only.')
+      return
+    }
+    setErrorMsg('')
+    const scoreObj = createScoreObj(score, name, gameType)
+    console.log('scoreObj', scoreObj)
+    postScore(scoreObj)
+    setOpenPostScore(false);
+    setGameOver(true)
+    // TODO: Implement Toast message
+  }
+
+  const postScore = async (scoreObj) => {
+    console.log('scoreObj str', JSON.stringify(scoreObj))
+    const strJSON = JSON.stringify(scoreObj)
+    try {
+      const response = await fetch('https://7zmgqfw2d1.execute-api.us-west-1.amazonaws.com/score',{
+        method: 'POST',
+        headers: {
+          Accept: 'application/json',
+        },
+        body: strJSON,
+      });
+      const json = await response.json()
+      console.log("response". json)
+      return json;
+    } catch (error) {
+      console.error(error);
+    }
+  }
+
   return (
     <View style={styles.container}>
       {gameOver ? <GameResultComponent score={score} handlePlayAgain={handlePlayAgain} db={db} gameType={gameType} navigation={navigation} savedScore={savedScore} /> :
@@ -161,6 +234,23 @@ export default function GameComponent({ navigation, gameType, startTime = 30, db
         </View> :
           <GameStartComponent handleStart={handleStart} gameType={gameType} startTime={startTime} navigation={navigation} />
       }
+      <Modal visible={openPostScore} contentContainerStyle={styles.postScoreContainer} onDismiss={() => navigation.navigate('Home')}>
+        <View style={styles.modalTitleRow}>
+          <Text style={styles.modalTitle}>Post Score</Text>
+        </View>
+        <View style={styles.formContainer}>
+          <View style={[styles.formElement, styles.nameInputContainer]}>
+            <TextInput mode='outlined' label='Name' value={name} onChangeText={(text) => setName(text)} error={errorMsg.length > 0}></TextInput>
+            {errorMsg.length > 0 ? <HelperText type="error" >
+              {errorMsg}
+            </HelperText> : null}
+            <TextInput style={{marginVertical: '5%'}} disabled={true} mode='outlined' label='Score' value={score > 0 ? score : ' '}>{score}</TextInput>
+          </View>
+          <TouchableOpacity style={styles.buttonStyle} onPress={() => handlePostScore()}>
+            <Text style={styles.buttonLabel}>Post Score</Text>
+          </TouchableOpacity>
+        </View>
+      </Modal>
     </View>
   );
 }
